pyproject.toml
[project]
name = "jwt-auth-fastapi-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "passlib>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.31",
]
